Add files via upload
An image search engine is developed in Matlab. 

The type of images used in this academic research project are Satellite images of Tropical Cyclone. This work is a prototype to understand that a real time operational tropical cyclone forecasting system can be developed, by the understanding of the features that characterizes tropical cyclones in satellite images.